requirements.txt
langchain
langchain-openai
langchain-core
langchain_community
docx2txt
pypdf
langchain_chroma
python-multipart
python-dotenv
fastapi
uvicorn
sentence_transformers
langchain-huggingface

# openai faiss-cpu rank_bm25 flashrank langchain_groq groq
# %pip install -qU langchain-groq
# %pip install -qU langchain-community
# %pip install -qU sentence-transformers